Split reducers into seperate files, can still split them up further
